LPS-65744 Move util classes to internal package
