+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/AlloyTaglibDefaultsPlugin.java
@@ -16,7 +16,7 @@ package com.liferay.gradle.plugins;
 
 import com.liferay.gradle.plugins.alloy.taglib.AlloyTaglibPlugin;
 import com.liferay.gradle.plugins.alloy.taglib.BuildTaglibsTask;
-import com.liferay.gradle.plugins.util.GradleUtil;
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 
 import org.gradle.api.Action;
 import org.gradle.api.Project;
