+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/JSTranspilerDefaultsPlugin.java
@@ -14,9 +14,9 @@
 
 package com.liferay.gradle.plugins;
 
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.js.transpiler.JSTranspilerExtension;
 import com.liferay.gradle.plugins.js.transpiler.JSTranspilerPlugin;
-import com.liferay.gradle.plugins.util.GradleUtil;
 
 import org.gradle.api.Project;
