+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/LangBuilderDefaultsPlugin.java
@@ -14,9 +14,9 @@
 
 package com.liferay.gradle.plugins;
 
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.lang.builder.BuildLangTask;
 import com.liferay.gradle.plugins.lang.builder.LangBuilderPlugin;
-import com.liferay.gradle.plugins.util.GradleUtil;
 
 import org.gradle.api.Action;
 import org.gradle.api.Project;
