+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/LiferayPlugin.java
@@ -14,8 +14,8 @@
 
 package com.liferay.gradle.plugins;
 
-import com.liferay.gradle.plugins.util.FileUtil;
-import com.liferay.gradle.plugins.util.GradleUtil;
+import com.liferay.gradle.plugins.internal.util.FileUtil;
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 
 import java.io.File;
 import java.io.IOException;
