+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/LiferaySettingsPlugin.java
@@ -14,7 +14,7 @@
 
 package com.liferay.gradle.plugins;
 
-import com.liferay.gradle.plugins.util.GradleUtil;
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.util.Validator;
 
 import java.io.File;
