+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/UpgradeTableBuilderDefaultsPlugin.java
@@ -14,9 +14,9 @@
 
 package com.liferay.gradle.plugins;
 
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.upgrade.table.builder.BuildUpgradeTableTask;
 import com.liferay.gradle.plugins.upgrade.table.builder.UpgradeTableBuilderPlugin;
-import com.liferay.gradle.plugins.util.GradleUtil;
 
 import java.io.File;
