+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/WhipDefaultsPlugin.java
@@ -14,7 +14,7 @@
 
 package com.liferay.gradle.plugins;
 
-import com.liferay.gradle.plugins.util.GradleUtil;
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.plugins.util.PortalTools;
 import com.liferay.gradle.plugins.whip.WhipExtension;
 import com.liferay.gradle.plugins.whip.WhipPlugin;
