+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/extensions/AppServer.java
@@ -14,7 +14,7 @@
 
 package com.liferay.gradle.plugins.extensions;
 
-import com.liferay.gradle.plugins.util.GradleUtil;
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 import com.liferay.gradle.util.OSDetector;
 
 import java.io.File;
