+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/tasks/BuildThumbnailsTask.java
@@ -15,8 +15,8 @@
 package com.liferay.gradle.plugins.tasks;
 
 import com.liferay.gradle.plugins.LiferayOSGiPlugin;
-import com.liferay.gradle.plugins.util.FileUtil;
-import com.liferay.gradle.plugins.util.GradleUtil;
+import com.liferay.gradle.plugins.internal.util.FileUtil;
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
 
 import java.io.File;
 import java.io.OutputStream;
