+++ modules/sdk/gradle-plugins/src/main/java/com/liferay/gradle/plugins/util/PortalTools.java
@@ -14,6 +14,8 @@
 
 package com.liferay.gradle.plugins.util;
 
+import com.liferay.gradle.plugins.internal.util.GradleUtil;
+
 import java.io.IOException;
 import java.io.InputStream;
